requirements.txt
streamlit
pandas
openpyxl
python-dotenv
langchain
langchain-openai
langchain-community
langchain-experimental
sqlalchemy
